% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acceptable.R
\docType{data}
\name{acceptable_spellings}
\alias{acceptable_spellings}
\title{Acceptable Spellings}
\format{A data frame with 257 rows and 2 variables:
\describe{
  \item{target}{The target word (i.e., the word that should be used)}
  \item{acceptable}{The acceptable spelling of the target word.}
}}
\usage{
acceptable_spellings
}
\description{
A dataset containing homophones.
}
\keyword{datasets}
