% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_misspell.R
\docType{data}
\name{alternate_df_default}
\alias{alternate_df_default}
\title{Common Misspellings List}
\format{A data frame with 4,278 rows and 2 variables:
\describe{
  \item{original}{The original, generally-accepted spelling}
  \item{alternate}{The misspelling of the original word.}
}}
\usage{
alternate_df_default
}
\description{
A dataset containing common misspellings in the english lanugage.
}
\keyword{datasets}
