% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore.R
\name{assign_score}
\alias{assign_score}
\title{Internal Function: Automatically assign scores to each subjects given new data set and scoring table (Used for intermediate and final evaluation)}
\usage{
assign_score(df, score_table)
}
\arguments{
\item{df}{A \code{data.frame} used for testing, where variables keep before categorization}

\item{score_table}{A \code{vector} containing the scoring table}
}
\value{
Processed \code{data.frame} with assigned scores for each variables
}
\description{
Internal Function: Automatically assign scores to each subjects given new data set and scoring table (Used for intermediate and final evaluation)
}
