% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AutoScore}
\alias{AutoScore}
\title{Automatic Clinical Score Generation}
\usage{
AutoScore(MD3, n)
}
\arguments{
\item{MD3}{a dataframe that contains data to be analysed}

\item{n}{number of variables to select}
}
\value{
List of parameters generated by autoscore function
}
\description{
Automatic Clinical Score Generation
}
\examples{
AutoScore(data, n = 10)
}
