
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AutoScore_fine_tuning}
\alias{AutoScore_fine_tuning}
\title{Pepline function: STEP (4): Fine-tune the score(AutoScore Module 5)}
\usage{
AutoScore_fine_tuning(TrainSet, ValidationSet, FinalVariable, CutVec, MaxScore=100)
}
\arguments{
\item{TrainSet}{a dataframe that is Training set }

\item{ValidationSet}{a dataframe that is Validation Set}

\item{FinalVariable}{Final list of variables, generated from last step}

\item{MaxScore}{Predefined cap of final score, e.g. 100}

\item{CutVec}{Generated from STEP(3) \code{AutoScore_weighting().Please follow the guidebook}}

}
\value{
Generated final table of scoring model for downstream testing process[STEP (5)]
}
\description{
STEP (4): Fine-tune the score
- Revise CutVec with domain knowledge to update scoring table (AutoScore Module 5)
- Rerun AutoScore Modules 2+3
- User can choose any cut-off values/any number of categories
}
\examples{
ScoringTable <- AutoScore_fine_tuning(TrainSet, ValidationSet, FinalVariable, CutVec, MaxScore=100)
}
