% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{AutoScore_impute}
\alias{AutoScore_impute}
\title{Internal function: impute missing values in the training and validation sets}
\usage{
AutoScore_impute(train_set, validation_set = NULL)
}
\arguments{
\item{train_set}{A data.frame of the training data.}

\item{validation_set}{A data.frame of the validation data. Default is \code{NULL}.}
}
\value{
Returns the imputed sets.
}
\description{
Internal function: impute missing values in the training and validation sets
}
