% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AutoScore_insample}
\alias{AutoScore_insample}
\title{Direct Automatic Clinical Score Generation (using in-sample validation)}
\usage{
AutoScore_insample(data, m, MaxScore = 100, probs = c(0, 0.05, 0.2, 0.8, 0.95, 1))
}
\arguments{
\item{data}{a dataframe that contains data to be analysed}

\item{m}{Predefined number of variables to be selected}

\item{MaxScore}{Predefined cap of final score, e.g. 100}

\item{probs}{Predefine quantiles to convert continuous variables to categorical, default:(0, 0.05, 0.2, 0.8, 0.95, 1)}
}
\value{
List of parameters generated by model: including list of variable selected, scoring table and performance evaluation (in-sample)
}
\description{
This function is used to generate scoring model based on a dataset and predefined number of variables.
And it uses all sample for performance evaluation, which is good for some studied with small sample size.
}
\examples{
AutoScore_insample(data=Sample_Data, m=8))
}
