% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AutoScore_outofsample}
\alias{AutoScore_outofsample}
\title{Direct Automatic Clinical Score Generation (using out-of-sample validation)}
\usage{
AutoScore_outofsample(Dataset, m = 8 , Percentage_test = 0.2, MaxScore = 100, probs = c(0, 0.05, 0.2, 0.8, 0.95, 1))
}
\arguments{
\item{Dataset}{a dataframe that contains data to be analysed}

\item{m}{Predefined number of variables to be selected}

\item{Percentage_test}{percentage of samples used for testing(out-of-sample validation), e.g. 0.2}

\item{MaxScore}{Predefined cap of final score, e.g. 100}

\item{probs}{Predefine quantiles to convert continuous variables to categorical, default:(0, 0.05, 0.2, 0.8, 0.95, 1)}
}
\value{
List of parameters generated by model: including list of variable selected, scoring table and performance evaluation (out-of-sample)
}
\description{
This function is used to generate scoring model based on a dataset and predefined number of variables.
And it uses testing data randomly selected from the original dataset for performance evaluation, which is good for some studied with relatively big sample size.(if sample size is small, please use \code{AutoScore_insample()})
}
\examples{
AutoScore_outofsample(Sample_Data, m=8))
}
