
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AutoScore_parsimony}
\alias{AutoScore_parsimony}
\title{Pepline function: STEP (2):Select the best model with parsimony plot (AutoScore Modules 2+3+4)}
\usage{
AutoScore_parsimony(TrainSet, ValidationSet, rank=Ranking, nmin=1, nmax=20, probs=c(0, 0.05, 0.2, 0.8, 0.95, 1))
}
\arguments{
\item{TrainSet}{a dataframe that is Training set }

\item{ValidationSet}{a dataframe that is Validation Set}

\item{rank}{the result generated from STEP (1)}

\item{nmin}{Minimum number of selected variables, default:1}

\item{nmax}{Maximum number of selected variables, default:20}

\item{probs}{ Predefine quantiles to convert continuous variables to categorical, default:(0, 0.05, 0.2, 0.8, 0.95, 1)}

}
\value{
List of AUC value for different parameter m and parsimony plot
}
\description{
STEP (2):Select the best model with parsimony plot (AutoScore Modules 2+3+4)
}
\examples{
AUC <- AutoScore_parsimony(TrainSet, ValidationSet, rank=Ranking, nmin=1, nmax=20, probs=c(0, 0.05, 0.2, 0.8, 0.95, 1))

}
