% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AutoScore_rank}
\alias{AutoScore_rank}
\title{Pepline function: STEP (1):Genrate variable ranking List by machine learning (AutoScore Module 1)}
\usage{
AutoScore_rank(TrainSet, ntree=100)
}
\arguments{
\item{TrainSet}{a dataframe that contains data to be analysed, for training}

\item{ntree}{Number of trees in random forest algorithm, default:100}

}
\value{
List of variables and its ranking generated by machine learning(random forest)
}
\description{
STEP (1): Genrate variable ranking List (AutoScore Module 1)
}
\examples{
Ranking <- AutoScore_rank(TrainSet, ntree=200)
}
