
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AutoScore_testing}
\alias{AutoScore_testing}
\title{Pepline function: STEP (5): Final score evaluation (AutoScore Module 6)}
\usage{
AutoScore_testing(TestSet, FinalVariable, CutVec, ScoringTable)
}
\arguments{
\item{TestSet}{a dataframe that is Testing set }

\item{FinalVariable}{Final list of variables, generated from last step}

\item{ScoringTable}{Generated from STEP(4) \code{AutoScore_fine_tuning}.Please follow the guidebook}

\item{CutVec}{Generated from STEP(3) \code{AutoScore_weighting().Please follow the guidebook}

}
\value{
No return value. All performance evaluation would be printed out.
}
\description{
STEP (5): Final score evaluation (AutoScore Module 6)
}
\examples{
AutoScore_testing(TestSet, FinalVariable, CutVec, ScoringTable)
}
