% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AutoScore_validation}
\alias{AutoScore_validation}
\title{Automating Clinical Score Generation (Validation)}
\usage{
AutoScore_validation(Dataset, n)
}
\arguments{
\item{Dataset}{A dataframe that contains data to be analysed}

\item{n}{Number of variables to select}
}
\value{
List of parameters generated by model built
}
\description{
Automating Clinical Score Generation (Validation)
}
\examples{
AutoScore_validation <- function(data, n = 12)
}
