% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AutoScore_validation_range}
\alias{AutoScore_validation_range}
\title{Determine Validation Range Function}
\usage{
AutoScore_validation_range(Dataset, nmin = 1, nmax = 20)
}
\arguments{
\item{Dataset}{A dataframe that contains data to be analysed}

\item{nmin}{Minimum number of variables}

\item{nmax}{Maximum number of variables}
}
\value{
List of AutoScore-generated Validations
}
\description{
Determine Validation Range Function
}
\examples{
AutoScore_validation_range(data, nmin = 1, nmax = 20)
}
