
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AutoScore_weighting}
\alias{AutoScore_weighting}
\title{Pepline function: STEP (3): Generate initial score with Final Variable list (Rerun AutoScore Module 2+3)}
\usage{
AutoScore_weighting(TrainSet, ValidationSet, FinalVariable, MaxScore=100, probs=c(0, 0.05, 0.2, 0.8, 0.95, 1)
}
\arguments{
\item{TrainSet}{a dataframe that is Training set }

\item{ValidationSet}{a dataframe that is Validation Set}

\item{FinalVariable}{Final list of variables, generated from last step}

\item{MaxScore}{Predefined cap of final score, e.g. 100}

\item{probs}{ Predefine quantiles to convert continuous variables to categorical, default:(0, 0.05, 0.2, 0.8, 0.95, 1)}

}
\value{
Generated \code{CutVec} for downstream fine-tuning process[STEP (4)]
}
\description{
STEP (3): Generate initial score with Final Variable list (Rerun AutoScore Module 2+3)
}
\examples{
CutVec <- AutoScore_weighting(TrainSet, ValidationSet, FinalVariable, MaxScore=100, probs=c(0, 0.05, 0.2, 0.8, 0.95, 1))
}
