% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Ordinal.R
\name{check_data_ordinal}
\alias{check_data_ordinal}
\title{AutoScore function for ordinal outcomes: Check whether the input
  dataset fulfil the requirement of the AutoScore}
\usage{
check_data_ordinal(data)
}
\arguments{
\item{data}{The data to be checked}
}
\value{
No return value, the result of the checking will be printed out.
}
\description{
AutoScore function for ordinal outcomes: Check whether the input
  dataset fulfil the requirement of the AutoScore
}
\examples{
data("sample_data_ordinal")
check_data_ordinal(sample_data_ordinal)
}
