% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Survival.R
\name{check_data_survival}
\alias{check_data_survival}
\title{AutoScore function for survival data: Check whether the input
  dataset fulfill the requirement of the AutoScore}
\usage{
check_data_survival(data)
}
\arguments{
\item{data}{The data to be checked}
}
\value{
No return value, the result of the checking will be printed out.
}
\description{
AutoScore function for survival data: Check whether the input
  dataset fulfill the requirement of the AutoScore
}
\examples{
data("sample_data_survival")
check_data_survival(sample_data_survival)
}
