% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Ordinal.R
\name{check_link}
\alias{check_link}
\title{Internal function: Check link function}
\usage{
check_link(link)
}
\arguments{
\item{link}{the link function used to model ordinal outcomes. Default is
\code{"logit"} for proportional odds model. Other options are
\code{"cloglog"} (proportional hazards model) and \code{"probit"}.}
}
\description{
Internal function: Check link function
}
