% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{check_predictor}
\alias{check_predictor}
\title{Internal function: Check predictors}
\usage{
check_predictor(data_predictor)
}
\arguments{
\item{data_predictor}{Predictors to be checked}
}
\value{
No return value, the result of the checking will be printed out.
}
\description{
Internal function: Check predictors
}
