% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{compound_fixer}
\alias{compound_fixer}
\title{compound_fixer}
\usage{
compound_fixer(response, comp)
}
\arguments{
\item{response}{the response variable}

\item{comp}{a named vector of compounds (e.g., \code{c("rose bush" = "rosebush")} where the response is "rosebush" and the desired change is "rose bush").}
}
\description{
This function uses a default list of compound words (common to the SIT) and replaces the response where appropriate to match the right usage.
}
\examples{

response = c("rose bush", "junkyard", "door man", "cracker jack")
compound_fixer(
  response,
  comp = c(
    "junk yard" = "junkyard",
    "crackerjack" = "cracker jack"
))

}
