% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Survival.R
\name{compute_auc_val_survival}
\alias{compute_auc_val_survival}
\title{Internal function for survival outcomes: Compute AUC based on validation set for plotting parsimony (AutoScore Module 4)}
\usage{
compute_auc_val_survival(
  train_set_1,
  validation_set_1,
  variable_list,
  categorize,
  quantiles,
  max_cluster,
  max_score
)
}
\arguments{
\item{train_set_1}{Processed training set}

\item{validation_set_1}{Processed validation set}

\item{variable_list}{List of included variables}

\item{categorize}{Methods for categorize continuous variables. Options include "quantile" or "kmeans"}

\item{quantiles}{Predefined quantiles to convert continuous variables to categorical ones. Available if \code{categorize = "quantile"}.}

\item{max_cluster}{The max number of cluster (Default: 5). Available if \code{categorize = "kmeans"}.}

\item{max_score}{Maximum total score}
}
\value{
A List of AUC for parsimony plot
}
\description{
Compute AUC based on validation set for plotting parsimony (survival outcomes)
}
