% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Ordinal.R
\name{compute_final_score_ord}
\alias{compute_final_score_ord}
\title{Internal function: Compute risk scores for ordinal data given variables
selected, cut-off values and scoring table}
\usage{
compute_final_score_ord(data, final_variables, cut_vec, scoring_table)
}
\arguments{
\item{data}{A processed \code{data.frame} that contains data for validation
or testing purpose. This \code{data.frame} must have variable \code{label}
and should have same format as \code{train_set} (same variable names and
outcomes)}

\item{final_variables}{A vector containing the list of selected variables,
selected from Step(ii) \code{\link{AutoScore_parsimony_Ordinal}}.}

\item{cut_vec}{Generated from STEP(iii) \code{\link{AutoScore_weighting_Ordinal}}.}

\item{scoring_table}{The final scoring table after fine-tuning, generated
from STEP(iv) \code{\link{AutoScore_fine_tuning_Ordinal}}.Please follow the
guidebook}
}
\description{
Internal function: Compute risk scores for ordinal data given variables
selected, cut-off values and scoring table
}
