% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore.R
\name{compute_multi_variable_table}
\alias{compute_multi_variable_table}
\title{AutoScore function: Multivariate Analysis}
\usage{
compute_multi_variable_table(df)
}
\arguments{
\item{df}{data frame after checking}
}
\value{
result of the multivariate analysis
}
\description{
Generate tables for multivariate analysis
}
\examples{
data("sample_data")
names(sample_data)[names(sample_data) == "Mortality_inpatient"] <- "label"
multi_table<-compute_multi_variable_table(sample_data)
}
