% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Survival.R
\name{compute_multi_variable_table_survival}
\alias{compute_multi_variable_table_survival}
\title{AutoScore function for survival outcomes: Multivariate Analysis}
\usage{
compute_multi_variable_table_survival(df)
}
\arguments{
\item{df}{data frame after checking}
}
\value{
result of the multivariate analysis for survival outcomes
}
\description{
Generate tables for multivariate analysis for survival outcomes
}
\examples{
data("sample_data_survival")
multi_table<-compute_multi_variable_table_survival(sample_data_survival)
}
