% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Survival.R
\name{compute_score_table_survival}
\alias{compute_score_table_survival}
\title{Internal function: Compute scoring table for survival outcomes based on training dataset}
\usage{
compute_score_table_survival(train_set_2, max_score, variable_list)
}
\arguments{
\item{train_set_2}{Processed training set after variable transformation (AutoScore Module 2)}

\item{max_score}{Maximum total score}

\item{variable_list}{List of included variables}
}
\value{
A scoring table
}
\description{
Compute scoring table for survival outcomes based on training dataset
}
