% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore.R
\name{compute_uni_variable_table}
\alias{compute_uni_variable_table}
\title{AutoScore function: Univariable Analysis}
\usage{
compute_uni_variable_table(df)
}
\arguments{
\item{df}{data frame after checking}
}
\value{
result of univariate analysis
}
\description{
Perform univariable analysis and generate the result table with odd ratios.
}
\examples{
data("sample_data")
names(sample_data)[names(sample_data) == "Mortality_inpatient"] <- "label"
uni_table<-compute_uni_variable_table(sample_data)
}
