% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Survival.R
\name{compute_uni_variable_table_survival}
\alias{compute_uni_variable_table_survival}
\title{AutoScore function for survival outcomes: Univariate Analysis}
\usage{
compute_uni_variable_table_survival(df)
}
\arguments{
\item{df}{data frame after checking}
}
\value{
result of the Univariate analysis for survival outcomes
}
\description{
Generate tables for Univariate analysis for survival outcomes
}
\examples{
data("sample_data_survival")
uni_table<-compute_uni_variable_table_survival(sample_data_survival)
}
