% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_contractions.R
\docType{data}
\name{contractions_df}
\alias{contractions_df}
\title{Common Contractions}
\format{
A data frame with 27 rows and 2 variables:
\describe{
  \item{contraction}{The contraction to replace}
  \item{replacement}{The replacement for the contraction}
}
}
\usage{
contractions_df
}
\description{
A dataset containing contractions
}
\keyword{datasets}
