% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore.R
\name{conversion_table}
\alias{conversion_table}
\title{AutoScore function: Print conversion table based on final performance evaluation}
\usage{
conversion_table(
  pred_score,
  by = "risk",
  values = c(0.01, 0.05, 0.1, 0.2, 0.5)
)
}
\arguments{
\item{pred_score}{a vector with outcomes and final scores generated from \code{\link{AutoScore_fine_tuning}}}

\item{by}{specify correct method for categorizing the threshold:  by "risk" or "score".Default to "risk"}

\item{values}{A vector of threshold for analyze sensitivity, specificity and other metrics. Default to "c(0.01,0.05,0.1,0.2,0.5)"}
}
\value{
No return value and the conversion will be printed out directly.
}
\description{
Print conversion table based on final performance evaluation
}
\seealso{
\code{\link{AutoScore_testing}}
}
