% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Survival.R
\name{conversion_table_survival}
\alias{conversion_table_survival}
\title{AutoScore function for survival outcomes: Print conversion table}
\usage{
conversion_table_survival(
  pred_score,
  score_cut = c(40, 50, 60),
  time_point = c(7, 14, 30, 60, 90)
)
}
\arguments{
\item{pred_score}{a data frame with outcomes and final scores generated from \code{\link{AutoScore_testing_Survival}}}

\item{score_cut}{Score cut-offs to be used for generating conversion table}

\item{time_point}{The time points to be evaluated using time-dependent AUC(t).}
}
\value{
conversion table and the it will also be printed out directly.
}
\description{
Print conversion table for survival outcomes
}
\seealso{
\code{\link{AutoScore_testing_Survival}}
}
