% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Descriptive}
\alias{Descriptive}
\title{Descriptive}
\usage{
Descriptive(x)
}
\arguments{
\item{x}{dataset after preprocessing}
}
\value{
print Descriptive table for your data, straified by your label(outcome)
}
\description{
Table one
}
\examples{
Descriptive(data)
}
