% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_word_detect.R
\name{double_word_detect}
\alias{double_word_detect}
\title{double_word_detect}
\usage{
double_word_detect(target)
}
\arguments{
\item{target}{the target variable}
}
\description{
This function marks the sentences with words that are repeated.
}
