% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Ordinal.R
\name{estimate_p_mat}
\alias{estimate_p_mat}
\title{Internal function: generate probability matrix for ordinal outcomes given
thresholds, linear predictor and link function}
\usage{
estimate_p_mat(theta, z, link)
}
\arguments{
\item{theta}{numeric vector of thresholds}

\item{z}{numeric vector of linear predictor}

\item{link}{The link function used to model ordinal outcomes. Default is
\code{"logit"} for proportional odds model. Other options are
\code{"cloglog"} (proportional hazards model) and \code{"probit"}.}
}
\description{
Internal function: generate probability matrix for ordinal outcomes given
thresholds, linear predictor and link function
}
