% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Survival.R
\name{eva_performance_iauc}
\alias{eva_performance_iauc}
\title{calculate iAUC for validation set (survival outcome)}
\usage{
eva_performance_iauc(score, ValidationSet, print = TRUE)
}
\description{
calculate iAUC for validation set (survival outcome)
}
