% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Ordinal.R
\name{evaluate_model_ord}
\alias{evaluate_model_ord}
\title{Internal function: Evaluate model performance on ordinal data}
\usage{
evaluate_model_ord(label, score, n_boot, report_cindex = TRUE)
}
\arguments{
\item{label}{outcome variable}

\item{score}{predicted score}

\item{n_boot}{Number of bootstrap cycles to compute 95%% CI for performance
metrics.}

\item{report_cindex}{If generalized c-index should be reported alongside
mAUC (Default:FALSE).}
}
\value{
Returns a list of the mAUC (mauc) and generalized c-index (cindex, if
  requested for) and their 95% bootstrap CIs (mauc_ci and cindex_ci).
}
\description{
Internal function: Evaluate model performance on ordinal data
}
