% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example Data Set}
\format{A data frame with 40 rows and 4 variables:
\describe{
  \item{Id}{ID}
  \item{Target}{The target phrase.}
  \item{Response}{The response phrase.}
  \item{human}{The human judgement of the number of matching words}
}}
\usage{
example_data
}
\description{
A dataset containing information on the results of a word identification study in speech perception research.
It has information on the target, response, and the human judgement of the accuracy of the word identification.
This is used in the README file and can be used as a reference for how data should be formatted.
}
\keyword{datasets}
