% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{find_one_inds}
\alias{find_one_inds}
\title{Internal function: Find column indices in design matrix that should be 1}
\usage{
find_one_inds(x_inds)
}
\arguments{
\item{x_inds}{A list of column indices corresponding to each final variable.}
}
\description{
Internal function: Find column indices in design matrix that should be 1
}
