% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{find_possible_scores}
\alias{find_possible_scores}
\title{Internal function: Compute all scores attainable.}
\usage{
find_possible_scores(final_variables, scoring_table)
}
\arguments{
\item{final_variables}{A vector containing the list of selected variables.}

\item{scoring_table}{The final scoring table after fine-tuning.}
}
\value{
Returns a numeric vector of all scores attainable.
}
\description{
Internal function: Compute all scores attainable.
}
