% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore.R
\name{get_cut_vec}
\alias{get_cut_vec}
\title{Internal function: Calculate cut_vec from the training set (AutoScore Module 2)}
\usage{
get_cut_vec(
  df,
  quantiles = c(0, 0.05, 0.2, 0.8, 0.95, 1),
  max_cluster = 5,
  categorize = "quantile"
)
}
\arguments{
\item{df}{training set to be used for calculate the cut vector}

\item{quantiles}{Predefined quantiles to convert continuous variables to categorical ones. (Default: c(0, 0.05, 0.2, 0.8, 0.95, 1)) Available if \code{categorize = "quantile"}.}

\item{max_cluster}{The max number of cluster (Default: 5). Available if \code{categorize = "kmeans"}.}

\item{categorize}{Methods for categorize continuous variables. Options include "quantile" or "kmeans" (Default: "quantile").}
}
\value{
cut_vec for \code{transform_df_fixed}
}
\description{
Internal function: Calculate cut_vec from the training set (AutoScore Module 2)
}
