% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_homophones.R
\docType{data}
\name{homophones}
\alias{homophones}
\title{Homophones}
\format{A data frame with 479 rows and 2 variables:
\describe{
  \item{rowname}{The row number.}
  \item{homophone}{The homophones (each line contains two or more words separated by commas).}
}}
\source{
\url{http://homophonelist.com/homophones-list/}
}
\usage{
homophones
}
\description{
A dataset containing homophones.
}
\keyword{datasets}
