% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{induce_informative_missing}
\alias{induce_informative_missing}
\title{Internal function: induce informative missing to sample data in the package
to demonstrate how AutoScore handles missing as a separate category}
\usage{
induce_informative_missing(
  df,
  vars_to_induce = c("Lab_A", "Vital_A"),
  prop_missing = 0.4
)
}
\arguments{
\item{df}{A data.frame of sample data.}

\item{vars_to_induce}{Names of variables to induce informative missing in.
Default is c("Lab_A", "Vital_A").}

\item{prop_missing}{Proportion of missing to induce for each
\code{vars_to_induce}. Can be a single value for a common proportion for
all variables (default is 0.4), or a vector with same length as
\code{vars_to_induce}.}
}
\value{
Returns \code{df} with selected columns modified to have missing.
}
\description{
Internal function: induce informative missing to sample data in the package
to demonstrate how AutoScore handles missing as a separate category
}
\details{
Assume subjects with normal values (i.e., values close to the
  median) are more likely to not have measurements.
}
