% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{induce_median_missing}
\alias{induce_median_missing}
\title{Internal function: induce informative missing in a single variable}
\usage{
induce_median_missing(x, prop_missing)
}
\arguments{
\item{x}{Variable to induce missing in.}

\item{prop_missing}{Proportion of missing to induce for each
\code{vars_to_induce}. Can be a single value for a common proportion for
all variables (default is 0.4), or a vector with same length as
\code{vars_to_induce}.}
}
\description{
Internal function: induce informative missing in a single variable
}
