% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Ordinal.R
\name{inv_cloglog}
\alias{inv_cloglog}
\title{Internal function: Inverse cloglog link}
\usage{
inv_cloglog(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Internal function: Inverse cloglog link
}
