% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Ordinal.R
\name{inv_probit}
\alias{inv_probit}
\title{Internal function: Inverse probit link}
\usage{
inv_probit(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Internal function: Inverse probit link
}
