% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{lasso}
\alias{lasso}
\title{LASSO (least absolute shrinkage and selection operator) Function}
\usage{
lasso(MD3)
}
\arguments{
\item{x}{}
}
\value{
list of parameters generated by lasso
}
\description{
LASSO (least absolute shrinkage and selection operator) Function
}
\examples{
lasso(data)
}
