% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{make_design_mat}
\alias{make_design_mat}
\title{Internal function: Based on \code{find_one_inds}, make a design matrix to
compute all scores attainable.}
\usage{
make_design_mat(one_inds)
}
\arguments{
\item{one_inds}{Output from \code{find_one_inds}.}
}
\description{
Internal function: Based on \code{find_one_inds}, make a design matrix to
compute all scores attainable.
}
