% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MultiVariable}
\alias{MultiVariable}
\title{Multivariable Analysis}
\usage{
MultiVariable(x)
}
\arguments{
\item{x}{}
}
\value{
Dataframe of parameters of multivariable analysis
}
\description{
Multivariable Analysis
}
\examples{
MultiVariable(data)
}
