% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{plot_auc}
\alias{plot_auc}
\title{Internal function: Make parsimony plot}
\usage{
plot_auc(
  AUC,
  variables,
  num = seq_along(variables),
  auc_lim_min,
  auc_lim_max,
  ylab = "Mean Area Under the Curve",
  title = "Parsimony plot on the validation set"
)
}
\arguments{
\item{AUC}{A vector of AUC values (or mAUC for ordinal outcomes).}

\item{variables}{A vector of variable names}

\item{num}{A vector of indices for AUC values to plot. Default is to plot all.}

\item{auc_lim_min}{Min y_axis limit in the parsimony plot (Default: 0.5).}

\item{auc_lim_max}{Max y_axis limit in the parsimony plot (Default: "adaptive").}

\item{ylab}{Title of y-axis}

\item{title}{Plot title}
}
\description{
Internal function: Make parsimony plot
}
