% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{plot_importance}
\alias{plot_importance}
\title{Internal Function: Print plotted variable importance}
\usage{
plot_importance(ranking)
}
\arguments{
\item{ranking}{vector output generated by functions: AutoScore_rank, AutoScore_rank_Survival or AutoScore_rank_Ordinal}
}
\description{
Internal Function: Print plotted variable importance
}
\seealso{
\code{\link{AutoScore_rank}}, \code{\link{AutoScore_rank_Survival}}, \code{\link{AutoScore_rank_Ordinal}}
}
