% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore.R
\name{plot_roc_curve}
\alias{plot_roc_curve}
\title{Internal Function: Plotting ROC curve}
\usage{
plot_roc_curve(prob, labels, quiet = TRUE)
}
\arguments{
\item{prob}{Predicate probability}

\item{labels}{Actual outcome(binary)}

\item{quiet}{if set to TRUE, there will be no trace printing}
}
\value{
No return value and the ROC curve will be plotted.
}
\description{
Internal Function: Plotting ROC curve
}
