% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Preprocess}
\alias{Preprocess}
\title{Preprocess Dataset}
\usage{
Preprocess(data, outcome)
}
\arguments{
\item{data}{The data which will be the source for the dataset}

\item{outcome}{The column within the dataset that determines the outcome}
}
\value{
The preprocessed dataset
}
\description{
Preprocess Dataset
}
\examples{
data <- Preprocess(testdf1_mimic, outcome = "label")
}
