% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Ordinal.R
\name{print_performance_ordinal}
\alias{print_performance_ordinal}
\title{AutoScore function for ordinal outcomes: Print predictive performance}
\usage{
print_performance_ordinal(label, score, n_boot = 100, report_cindex = FALSE)
}
\arguments{
\item{label}{outcome variable}

\item{score}{predicted score}

\item{n_boot}{Number of bootstrap cycles to compute 95%% CI for performance
metrics.}

\item{report_cindex}{Whether to report generalized c-index for model
evaluation (Default:FALSE for faster evaluation).}
}
\value{
No return value and the ROC performance will be printed out directly.
}
\description{
Print mean area under the curve (mAUC) and generalised c-index
  (if requested)
}
\seealso{
\code{\link{AutoScore_testing_Ordinal}}
}
