% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Survival.R
\name{print_performance_survival}
\alias{print_performance_survival}
\title{AutoScore function for survival outcomes: Print predictive performance}
\usage{
print_performance_survival(score, validation_set, time_point)
}
\arguments{
\item{score}{Predicted score}

\item{validation_set}{Dataset for generating performance}

\item{time_point}{The time points to be evaluated using time-dependent AUC(t).}
}
\value{
No return value and the ROC performance will be printed out directly.
}
\description{
Print mean area under the curve (mAUC) and generalised c-index
  (if requested)
}
\seealso{
\code{\link{AutoScore_testing_Ordinal}}
}
