% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore.R
\name{print_score_table}
\alias{print_score_table}
\title{AutoScore Function: print_score_table}
\usage{
print_score_table(scoring_table, final_variable)
}
\arguments{
\item{scoring_table}{Raw scoring table generated by procedure}

\item{final_variable}{Final included variables by predictor_rank}
}
\value{
dataframe of formatted scoring table
}
\description{
Print scoring tables for visulization
}
\examples{
\dontrun{
print_score_table(scoring_table,final_variable)}
}
