% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwc.R
\name{pwc}
\alias{pwc}
\title{Percent Words Correct}
\usage{
pwc(data, id)
}
\arguments{
\item{data}{data.frame containing the target, response, and autoscore columns as produced by the function `autoscore()`.}

\item{id}{the id variable to summarize the PWC for}
}
\value{
data.frame of the percent correct for each "id"
}
\description{
Obtain the percent words correct from a data.frame produced by `autoscore()`.
}
