% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{20000 simulated ICU admission data, with the same distribution as the data in the MIMIC-III ICU database}
\format{
An object of class \code{data.frame} with 20000 rows and 22 columns.
}
\usage{
sample_data
}
\description{
20000 simulated samples, with the same distribution as the data in the MIMIC-III ICU database. It is used for demostration only in the Guidebook.
}
\keyword{datasets}
