% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Ordinal.R
\docType{data}
\name{sample_data_ordinal}
\alias{sample_data_ordinal}
\title{Simulated ED data with ordinal outcome}
\format{
An object of class \code{data.frame} with 20000 rows and 21 columns.
}
\usage{
sample_data_ordinal
}
\description{
Simulated data for 20,000 inpatient visits with demographic
  information, healthcare resource utilisation and associated laboratory
  tests and vital signs measured in the emergency department (ED). Data were
  simulated based on the dataset analysed in the AutoScore-Ordinal paper, and
  only includes a subset of variables (with masked variable names) for the
  purpose of demonstrating the AutoScore framework for ordinal outcomes.
}
\keyword{datasets}
