% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Ordinal.R
\docType{data}
\name{sample_data_ordinal_small}
\alias{sample_data_ordinal_small}
\title{Simulated ED data with ordinal outcome (small sample size)}
\format{
An object of class \code{data.frame} with 5000 rows and 21 columns.
}
\usage{
sample_data_ordinal_small
}
\description{
5,000 observations randomly sampled from
  \code{\link{sample_data_ordinal}}. It is used for demonstration only in the
  Guidebook.
}
\keyword{datasets}
