% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore.R
\docType{data}
\name{sample_data_small}
\alias{sample_data_small}
\title{1000 simulated ICU admission data, with the same distribution as the data in the MIMIC-III ICU database}
\format{
An object of class \code{data.frame} with 1000 rows and 22 columns.
}
\usage{
sample_data_small
}
\description{
1000 simulated samples, with the same distribution as the data in the MIMIC-III ICU database. It is used for demonstration only in the Guidebook. Run \code{vignette("Guide_book", package = "AutoScore")} to see the guidebook or vignette.
 \itemize{
 \item{Johnson, A., Pollard, T., Shen, L. et al. MIMIC-III, a freely accessible critical care database. Sci Data 3, 160035 (2016).}
}
}
\keyword{datasets}
