% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Survival.R
\docType{data}
\name{sample_data_survival}
\alias{sample_data_survival}
\title{20000 simulated MIMIC sample data with survival outcomes}
\format{
An object of class \code{data.frame} with 20000 rows and 23 columns.
}
\usage{
sample_data_survival
}
\description{
20000 simulated samples, with the same distribution
as the data in the MIMIC-III ICU database. Data were simulated based on the dataset
  analysed in the AutoScore-Survival paper. It is used for demonstration
only in the Guidebook. Run \code{vignette("Guide_book", package = "AutoScore")}
to see the guidebook or vignette.
 \itemize{
 \item{Johnson, A., Pollard, T., Shen, L. et al. MIMIC-III, a freely accessible critical care database. Sci Data 3, 160035 (2016).}
}
}
\keyword{datasets}
