% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore.R
\docType{data}
\name{sample_data_with_missing}
\alias{sample_data_with_missing}
\title{20000 simulated ICU admission data with missing values}
\format{
An object of class \code{data.frame} with 20000 rows and 23 columns.
}
\usage{
sample_data_with_missing
}
\description{
20000 simulated samples with missing values, which can be used for demostrating AutoScore workflow dealing with missing values.
 \itemize{
 \item{Johnson, A., Pollard, T., Shen, L. et al. MIMIC-III, a freely accessible critical care database. Sci Data 3, 160035 (2016).}
}
}
\keyword{datasets}
