% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tableone}
\alias{tableone}
\title{Table one}
\usage{
tableone(x)
}
\arguments{
\item{x}{}
}
\value{
List of MD_table and MD_table_overall
}
\description{
Table one
}
\examples{
tableone(data)
tableone_tables <- reactive({tableone(table_data())})
}
