% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore.R
\name{transform_df_fixed}
\alias{transform_df_fixed}
\title{Internal function: Categorizing continuous variables based on cut_vec (AutoScore Module 2)}
\usage{
transform_df_fixed(df, cut_vec)
}
\arguments{
\item{df}{dataset(training, validation or testing) to be processed}

\item{cut_vec}{fixed cut vector}
}
\value{
Processed \code{data.frame} after categorizing based on fixed cut_vec
}
\description{
Internal function: Categorizing continuous variables based on cut_vec (AutoScore Module 2)
}
