% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{UniVariable}
\alias{UniVariable}
\title{Univariable Analysis}
\usage{
UniVariable(x)
}
\arguments{
\item{x}{}
}
\value{
dataframe of univariate analysis
}
\description{
Univariable Analysis
}
\examples{
UniVariable(data)
}
