% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscore.R, R/autoscore_fun.R
\docType{package}
\name{autoscore}
\alias{autoscore}
\alias{autoscore-package}
\alias{autoscore}
\title{AutoScore}
\usage{
autoscore(.data, ..., output = "text")
}
\arguments{
\item{.data}{The data.frame (or tbl_df) to be used to autoscore}

\item{...}{the list of rules to apply}

\item{output}{the output type for the autoscore table; current options are "text" ...}
}
\description{
Description of the package and how to use it...

Function stuff
}
\examples{
\dontrun{

library(autoscore)


}

}
