% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AutoScore_insample}
\alias{AutoScore_insample}
\title{Automatic Clinical Score Generation (Insample)}
\usage{
AutoScore_insample(MD3, n)
}
\arguments{
\item{MD3}{a dataframe that contains data to be analysed}

\item{n}{number of variables to select}
}
\value{
List of parameters generated by model built
}
\description{
Automatic Clinical Score Generation (Insample)
}
\examples{
AutoScore_insample(data, n = 10)
}
