% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore.R
\name{change_reference}
\alias{change_reference}
\title{Internal Function: Change Reference category after first-step logistic regression (part of AutoScore Module 3)}
\usage{
change_reference(df, coef_vec)
}
\arguments{
\item{df}{A \code{data.frame} used for logistic regression}

\item{coef_vec}{Generated from logistic regression}
}
\value{
Processed \code{data.frame} after changing reference category
}
\description{
Internal Function: Change Reference category after first-step logistic regression (part of AutoScore Module 3)
}
