% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore.R
\name{check_data}
\alias{check_data}
\title{AutoScore function: Check whether the input dataset fulfill the requirement of the AutoScore}
\usage{
check_data(data)
}
\arguments{
\item{data}{The data to be checked}
}
\value{
No return value, the result of the checking will be printed out.
}
\description{
AutoScore function: Check whether the input dataset fulfill the requirement of the AutoScore
}
\examples{
data("sample_data")
names(sample_data)[names(sample_data) == "Mortality_inpatient"] <- "label"
check_data(sample_data)
}
