% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore.R
\name{compute_descriptive_table}
\alias{compute_descriptive_table}
\title{AutoScore function: Descriptive Analysis}
\usage{
compute_descriptive_table(df)
}
\arguments{
\item{df}{data frame after checking and fulfilling the requirement of AutoScore}
}
\value{
No return value and the result of the descriptive analysis will be printed out.
}
\description{
Compute descriptive table (usually Table 1 in the medical literature) for the dataset.
}
\examples{
data("sample_data")
names(sample_data)[names(sample_data) == "Mortality_inpatient"] <- "label"
compute_descriptive_table(sample_data)
}
