% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Survival.R
\name{eva_performance_iauc}
\alias{eva_performance_iauc}
\title{Internal function survival outcome: Calculate iAUC for validation set}
\usage{
eva_performance_iauc(score, validation_set, print = TRUE)
}
\arguments{
\item{score}{Predicted score}

\item{validation_set}{Dataset for generating performance}

\item{print}{Whether to print out the final iAUC result}
}
\description{
Internal function survival outcome: Calculate iAUC for validation set
}
