% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Ordinal.R
\name{extract_or_ci_ord}
\alias{extract_or_ci_ord}
\title{Extract OR, CI and p-value from a proportional odds model}
\usage{
extract_or_ci_ord(model, n_digits = 3)
}
\arguments{
\item{model}{An ordinal regression model fitted using \code{\link[ordinal]{clm}}.}

\item{n_digits}{Number of digits to print for OR or exponentiated
coefficients (Default:3).}
}
\description{
Extract OR, CI and p-value from a proportional odds model
}
