% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MultiVariable}
\alias{MultiVariable}
\title{Generate tables for Multivariable Analysis}
\description{
    Generate tables for Multivariable Analysis}
\usage{
MultiVariable(data)
}
\arguments{
\item{data}{standard data frame}
}
\value{
Dataframe of results of multivariable analysis
}

\examples{
MultiVariable(data)
}
