% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore_Survival.R
\name{print_performance_ci_survival}
\alias{print_performance_ci_survival}
\title{AutoScore function for survival outcomes: Print predictive performance}
\usage{
print_performance_ci_survival(score, ValidationSet, time_point, cycle = 100)
}
\arguments{
\item{score}{predicted score}

\item{label}{outcome variable}
}
\value{
No return value and the ROC performance will be printed out directly.
}
\description{
Print mean area under the curve (mAUC) and generalised c-index
  (if requested)
}
\seealso{
\code{\link{AutoScore_testing_Ordinal}}
}
