% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore.R
\name{print_roc_performance}
\alias{print_roc_performance}
\title{Internal function: Print receiver operating characteristic (ROC) performance}
\usage{
print_roc_performance(label, score, threshold = "best")
}
\arguments{
\item{label}{outcome}

\item{score}{predicted score}

\item{threshold}{Threshold for analyze sensitivity, specificity and other metrics. Default to "best"}
}
\description{
Print receiver operating characteristic (ROC) performance
}
\examples{
\dontrun{
print_roc_performance<-function(label, score, threshold = "best")}
}
