% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoScore.R
\name{print_scoring_table}
\alias{print_scoring_table}
\title{AutoScore Function: Print scoring tables for visualization}
\usage{
print_scoring_table(scoring_table, final_variable)
}
\arguments{
\item{scoring_table}{Raw scoring table generated by AutoScore step(iv) \code{\link{AutoScore_fine_tuning}}}

\item{final_variable}{Final included variables}
}
\value{
Data frame of formatted scoring table
}
\description{
AutoScore Function: Print scoring tables for visualization
}
\seealso{
\code{\link{AutoScore_fine_tuning}}, \code{\link{AutoScore_weighting}}
}
