% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwc.R
\name{pwc}
\alias{pwc}
\title{Percent Words Correct}
\usage{
pwc(data)
}
\arguments{
\item{data}{data.frame containing the target, response, and autoscore columns as produced by the function `autoscore()`.}
}
\value{
data.frame of the percent correct for each line in `data`
}
\description{
Obtain the percent words correct from a data.frame produced by `autoscore()`.
}
